+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to
 - ✨(backend) add command to clean pending and deleted files
 - 🧱(helm) run clean files command as cronjob
 
+### Changed
+
+- ✨(summary) generalized stt api call #1420
+
 ## [1.21.0] - 2026-06-15
 
 ### Added